PR Review update
- Preferring 5.minutes over 300 and cleaner comparison
- Preferring exists? object blank? for server where clause.
