+++ app/controllers/manageiq/api/base_controller/logger.rb
@@ -13,7 +13,13 @@ module ManageIQ
 
         def log_api_auth
           return unless api_log_info?
-          log_request("Authentication", :type        => @auth_token.blank? ? "basic" : "token",
+          if @miq_token_hash
+            auth_type = "system"
+            log_request("System Auth", {:x_miq_token => request.headers['X-MIQ-Token']}.merge(@miq_token_hash))
+          else
+            auth_type = @auth_token.blank? ? "basic" : "token"
+          end
+          log_request("Authentication", :type        => auth_type,
                                         :token       => @auth_token,
                                         :x_miq_group => request.headers['X-MIQ-Group'],
                                         :user        => @auth_user)
